+++ app/src/main/java/com/example/tidy/Utils.kt
@@ -1,6 +1,5 @@
 package com.example.tidy
 
-import com.example.tidy.viewModels.BackupDto
 import com.google.gson.Gson
 import java.text.SimpleDateFormat
 import java.util.Calendar
